# api-docs/awsomeshop-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AWSomeShop 统一网关 API",
    "description": "4服务合并,路径含网关前缀。v1.2 含设计稿缺口补齐",
    "version": "1.2"
  },
  "servers": [
    {
      "url": "http://localhost:8088",
      "description": "Gateway(本地)"
    }
  ],
  "paths": {
...
    "/auth/api/admin/users/{id}/status": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "更新用户状态（启用/禁用）",
        "operationId": "updateStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Auth_UpdateUserStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Auth_ResultUserDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/product/api/internal/products/restore-stock": {
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "回补库存（内部，Saga 补偿调用）",
        "operationId": "restoreStock",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product_StockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Product_ResultProductDTO"
                }
              }
            }
          }
        }
      }
    },
    "/product/api/internal/products/deduct-stock": {
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "扣减库存（内部，兑换调用）",
        "operationId": "deductStock",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product_StockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Product_ResultProductDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/product/api/admin/products/{id}/stock": {
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "管理员调整库存 (入库IN/出库OUT) 并记录",
        "operationId": "adjustStock",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "changeType",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Product_ResultProductDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/product/api/admin/products/{id}/status": {
      "patch": {
        "tags": [
          "Product"
        ],
        "summary": "更新商品状态（上架/下架）",
        "operationId": "updateStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product_UpdateProductStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Product_ResultProductDTO"
                }
              }
            }
          }
        }
      }
    },
    "/product/api/admin/categories/{id}/status": {
      "patch": {
        "tags": [
          "Category"
        ],
        "summary": "更新类目状态（启用/禁用）",
        "operationId": "updateStatus_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product_UpdateCategoryStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Product_ResultCategoryDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/product/api/products/{id}/related": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "同类推荐 (PROD-8) — 同分类其他商品 TOP 6",
        "operationId": "related",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Product_ResultListProductDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/point/api/admin/point-rules/{id}/status": {
      "patch": {
        "tags": [
          "PointRule"
        ],
        "summary": "启用/禁用积分规则",
        "operationId": "updateStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Point_UpdatePointRuleStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Point_ResultPointRuleDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/order/api/admin/orders/{id}/status": {
      "put": {
        "tags": [
          "ExchangeRecord"
        ],
        "summary": "更新兑换记录状态",
        "operationId": "updateStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order_UpdateExchangeStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Order_ResultExchangeRecordDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/order/api/orders": {
      "get": {
        "tags": [
          "ExchangeRecord"
        ],
        "summary": "员工查询自己的兑换记录",
        "operationId": "listMine",
        "parameters": [
          {
            "name": "request",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Order_ListMyExchangeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Order_ResultPageResultExchangeRecordDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ExchangeRecord"
        ],
        "summary": "员工兑换下单",
        "operationId": "exchange",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order_ExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Order_ResultExchangeRecordDTO"
                }
              }
            }
          }
        }
      }
    },
    "/order/api/orders/{id}/confirm-receipt": {
      "post": {
        "tags": [
          "ExchangeRecord"
        ],
        "summary": "员工确认收货",
        "operationId": "confirmReceipt",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Order_ResultExchangeRecordDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/order/api/orders/{id}": {
      "get": {
        "tags": [
          "ExchangeRecord"
        ],
        "summary": "员工查询兑换记录详情",
        "operationId": "getMine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Order_ResultExchangeRecordDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Auth_ResultUserDTO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Auth_UserDTO"
          }
        }
      },
      "Auth_UserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "employeeId": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Auth_UpdateUserStatusRequest": {
        "required": [
          "status",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "Product_ProductDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "pointsPrice": {
            "type": "integer",
            "format": "int32"
          },
          "marketPrice": {
            "type": "number"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "soldCount": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subtitle": {
            "type": "string"
          },
          "deliveryMethod": {
            "type": "string"
          },
          "serviceGuarantee": {
            "type": "string"
          },
          "promotion": {
            "type": "string"
          },
          "colors": {
            "type": "string"
          },
          "specs": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Product_ResultProductDTO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Product_ProductDTO"
          }
        }
      },
...
      "Product_CategoryDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "parentId": {
            "type": "integer",
            "format": "int64"
          },
          "icon": {
            "type": "string"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "productCount": {
            "type": "integer",
            "format": "int64"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product_CategoryDTO"
            }
          }
        }
      },
      "Product_ResultCategoryDTO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Product_CategoryDTO"
          }
        }
      },
...
      "Product_StockRequest": {
        "required": [
          "productId",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Product_UpdateProductStatusRequest": {
        "required": [
          "id",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Product_UpdateCategoryStatusRequest": {
        "required": [
          "id",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Product_ResultListProductDTO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product_ProductDTO"
            }
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "Point_PointRuleDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "ruleType": {
            "type": "string"
          },
          "pointValueMin": {
            "type": "integer",
            "format": "int32"
          },
          "pointValueMax": {
            "type": "integer",
            "format": "int32"
          },
          "triggerCondition": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "grantMethod": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Point_ResultPointRuleDTO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Point_PointRuleDTO"
          }
        }
      },
...
      "Point_UpdatePointRuleStatusRequest": {
        "required": [
          "id",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Order_UpdateExchangeStatusRequest": {
        "required": [
          "id",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "trackingNumber": {
            "type": "string"
          },
          "carrier": {
            "type": "string"
          }
        }
      },
...
      "Order_ResultExchangeRecordDTO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Order_ExchangeRecordDTO"
          }
        }
      },
...
      "Order_ExchangeRequest": {
        "required": [
          "productId",
          "userId"
        ],
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "employeeName": {
            "type": "string"
          },
          "addressId": {
            "type": "integer",
            "format": "int64"
          },
          "idempotencyKey": {
            "type": "string"
          }
        }
      },
      "Order_ListMyExchangeRequest": {
        "required": [
          "userId"
        ],
        "type": "object",
        "properties": {
          "page": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "maximum": 100,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "keyword": {
            "type": "string"
          }
        }
      },
      "Order_PageResultExchangeRecordDTO": {
        "type": "object",
        "properties": {
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "pages": {
            "type": "integer",
            "format": "int64"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order_ExchangeRecordDTO"
            }
          }
        }
      },
      "Order_ResultPageResultExchangeRecordDTO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Order_PageResultExchangeRecordDTO"
          }
        }
      },
...
    }
  }
}